Move replay playback rate control to its own component
